requirements.txt
2to3==1.0
ApiDoc==1.4.0
celery==5.0.5
Click==7.0
coreapi==2.3.3
Django==2.0.13
django-allauth==0.39.1
django-celery-results==2.0.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.2.0
django-model-utils==3.1.2
django-redis==4.9.0
django-user-agents==0.4.0
djangorestframework==3.10.3
drf-dynamic-fields==0.3.1
drf-yasg==1.16.1
etcd3==0.12.0
Faker==1.0.5
gevent==1.4.0
greenlet==0.4.15
grpcio==1.29.0
grpcio-tools==1.29.0
gunicorn==20.0.4
Jinja2==2.11.2
flower
ldaptor==19.1.0

Mako==1.1.0
Markdown==3.1.1

Pillow==6.0.0

protobuf==3.12.2

pyflakes==2.1.1

pylint==2.3.1
pylint-django==2.0.8
pylint-plugin-utils==0.5
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
pytest==4.4.1
pytest-django==3.4.8
python-dateutil==2.8.0
python-etcd==0.4.5
python-gettext==4.0
python-ldap==3.2.0
python-slugify==3.0.2
python3-memcached==1.51
python3-openid==3.1.0
pytz==2019.1
PyYAML
redis==3.2.1
requests==2.21.0

Sphinx==2.0.1

sqlparse==0.3.0

supervisor==4.2.0
supervisor-multiprocesslogging==1.1
toml==0.10.0
tornado==5.1.1

Twisted==19.7.0

virtualenv==20.0.23
vthread==0.1.1

#M2Crypto==0.36
rsa==4.6
zeppelin-pyclient
objectdict
pypinyin
addict
hdfs
furl
xlrd